Pair each type's serializer and deserializer in one codec class
BufferCodec, ByteArrayCodec, InstantCodec, JsonArrayCodec, and
JsonObjectCodec each hold their type's Serializer and Deserializer, halving
the file count the module spreads its wire format across. compileJava uses
options.release so javac verifies API usage against the Java 17 API instead
of only emitting 17 bytecode.
